+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 -   refactor Home View for i18n [#1142](https://github.com/EthVM/EthVM/pull/1142)
 -   Added Portfolio View for i18n [#1155](https://github.com/EthVM/EthVM/pull/1155)
 -   Added setting View for i18n [#1158](https://github.com/EthVM/EthVM/pull/1158)
-
+-   Added Transaction details View for i18n [#1159](https://github.com/EthVM/EthVM/pull/1159)
 
 ### Feature
 - advertise with us page [#1156](https://github.com/EthVM/EthVM/pull/1156)